safari-534.53-branch/LayoutTests/platform/chromium-win/http/tests/misc/authentication-sent-to-redirect-expected.txt

CONSOLE MESSAGE: line 25: Uncaught TypeError: Object [object Object] has no method 'setHandlesAuthenticationChallenges'
https://bugs.webkit.org/show_bug.cgi?id=40138
This test loads a php script which demands http authentication, then uses it to redirect to another script using that shows what authentication headers were sent with the final request.
It does this once each for HTTP 301, 302, 303, and 307 redirects.
If not running under DRT, enter any credentials when asked.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------


--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
